requirements.txt
Flask==1.1.1
psycopg2-binary==2.8.3
pyyaml==5.1.1
python-dateutil==2.8.0
requests==2.22
flask-rest-api==0.16.1
marshmallow==2.19.5
redis==3.3.1